added serializers and generics views for models
