+++ rentalcar/rentalcar/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rentalCarSystem',
+    'rest_framework',
 ]
 
 MIDDLEWARE = [
