--- templates/index.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CarRental-Express</title>
-    <style>
-        .content{
-            /*border: 3px solid #a30b15;*/
-            margin: auto;
-            width: 1200px;
-            height: auto;
-            /*border-radius: 12px;*/
-            margin-top: 10%;
-        }
-        .title{
-            /* border: 3px solid black; */
-            text-align: center;
-        }
-
-        h2{
-            color: #a30b15;
-            font-weight: bold;
-            font-family: Verdana, Geneva, Tahoma, sans-serif;
-        }
-
-        .cars{
-            /* border: 3px solid blue; */
-            display: flex;
-            text-align: center;
-            margin: 30px auto 30px auto;
-        }
-        .car{
-            /* border: 3px solid yellow; */
-            padding: 25px 15px 25px 15px;
-            margin: auto;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="content">
-            <div class="title">
-                <h2>CarRental-Express</h1>
-            </div>
-            <div class="cars">
-                <div class="car">
-                    <img src="https://cdn.pixabay.com/photo/2021/09/06/20/50/ford-mustang-6602480_1280.png" width="200px" style="margin-bottom: 30px;">
-                </div>
-                <div class="car">
-                    <img src="https://cdn.pixabay.com/photo/2021/01/01/21/57/mclaren-5880138_1280.png" width="350px">
-                </div>
-                <div class="car">
-                    <img src="https://cdn.pixabay.com/photo/2019/05/22/16/57/car-4222013_1280.png" width="300px">
-                </div>
-                <div class="car">
-                    <img src="https://cdn.pixabay.com/photo/2020/12/28/15/01/chevrolet-5867459_1280.png" width="300px" style="margin-bottom: 30px;">
-                </div>
-            </div> 
-        </div>
-    </div>
-</body>
-</html>
