--- tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.
